Revert "Di setup"
