+++ app/src/main/AndroidManifest.xml
@@ -9,7 +9,6 @@
     <uses-permission android:name="android.permission.POST_NOTIFICATIONS" />
 
     <application
-        android:name=".PvnApp"
         android:allowBackup="true"
         android:icon="@drawable/vpn_icon"
         android:label="@string/app_name"
@@ -17,9 +16,9 @@
         android:theme="@style/Theme.FptnClient"
         tools:targetApi="33">
         <activity
-            android:name=".auth.ui.AuthActivity"
-            android:configChanges="uiMode"
-            android:exported="true">
+            android:name=".views.LoginActivity"
+            android:exported="true"
+            android:label="@string/app_name">
             <intent-filter>
                 <action android:name="android.intent.action.MAIN" />
                 <category android:name="android.intent.category.LAUNCHER" />
