+++ app/src/main/java/com/filantrop/pvnclient/services/CustomVpnConnection.java
@@ -35,6 +35,7 @@ import java.util.stream.Stream;
 import inet.ipaddr.IPAddress;
 import inet.ipaddr.IPAddressString;
 import lombok.Getter;
+import lombok.Setter;
 
 public class CustomVpnConnection extends Thread {
 
@@ -58,9 +59,11 @@ public class CustomVpnConnection extends Thread {
 
     private PendingIntent mConfigureIntent;
 
-    OnEstablishListener onEstablishListener;
+    @Setter
+    private OnEstablishListener onEstablishListener;
 
-    public Instant connectionTime;
+    @Getter
+    private Instant connectionTime;
 
     private final ScheduledExecutorService scheduler = Executors.newSingleThreadScheduledExecutor();
     private final DataRateCalculator downloadRate = new DataRateCalculator(1000);
@@ -194,15 +197,15 @@ public class CustomVpnConnection extends Thread {
     }
 
     private void sendErrorMessageToUI(String msg) {
-        service.mHandler.sendMessage(Message.obtain(null, HandlerMessageTypes.ERROR.value, 0, 0, msg));
+        service.getMHandler().sendMessage(Message.obtain(null, HandlerMessageTypes.ERROR.getValue(), 0, 0, msg));
     }
 
     private void sendSpeedInfoToUI(String downloadSpeed, String uploadSpeed) {
-        service.mHandler.sendMessage(Message.obtain(null, HandlerMessageTypes.SPEED_INFO.value, 0, 0, Pair.create(downloadSpeed, uploadSpeed)));
+        service.getMHandler().sendMessage(Message.obtain(null, HandlerMessageTypes.SPEED_INFO.getValue(), 0, 0, Pair.create(downloadSpeed, uploadSpeed)));
     }
 
     private void sendConnectionStateToUI(ConnectionState connectionState) {
-        service.mHandler.sendMessage(Message.obtain(null, HandlerMessageTypes.CONNECTION_STATE.value, 0, 0, Pair.create(connectionState, Instant.now())));
+        service.getMHandler().sendMessage(Message.obtain(null, HandlerMessageTypes.CONNECTION_STATE.getValue(), 0, 0, Pair.create(connectionState, Instant.now())));
     }
 
     private String getTag() {
