+++ app/src/main/java/com/filantrop/pvnclient/views/SettingsActivity.java
@@ -23,6 +23,7 @@ import lombok.Getter;
 
 
 public class SettingsActivity extends AppCompatActivity {
+    private final String TAG = this.getClass().getName();
 
     private ListView serverListView;
     @Getter
@@ -66,7 +67,7 @@ public class SettingsActivity extends AppCompatActivity {
         });
 
         fptnViewModel = new ViewModelProvider(this).get(FptnServerViewModel.class);
-        fptnViewModel.serverDtoListLiveData.observe(this, fptnServerDtos -> {
+        fptnViewModel.getServerDtoListLiveData().observe(this, fptnServerDtos -> {
             if (fptnServerDtos != null && !fptnServerDtos.isEmpty()) {
                 serverListView.setAdapter(new FptnServerAdapter(fptnServerDtos, R.layout.settings_server_list_item)); // NEED TO CHANGE THE ITEM LAYOUT
                 setListViewHeightBasedOnChildren(serverListView);
