--- app/src/test/kotlin/com/filantrop/pvnclient/di/CheckModulesTest.kt
@@ -1,13 +1,0 @@
-package com.filantrop.pvnclient.di
-
-import org.junit.Test
-import org.junit.experimental.categories.Category
-import org.koin.test.verify.verify
-
-@Category(CheckModulesTest::class)
-class CheckModulesTest {
-    @Test
-    fun checkAllModules() {
-        appModule.verify()
-    }
-}
