--- auth/data/src/main/kotlin/com/filantrop/pvnclient/auth/data/di/AuthDataModule.kt
@@ -1,10 +1,0 @@
-package com.filantrop.pvnclient.auth.data.di
-
-import com.filantrop.pvnclient.auth.data.AuthRepositoryImpl
-import com.filantrop.pvnclient.auth.domain.AuthRepository
-import org.koin.dsl.module
-
-val authDataModule =
-    module {
-        single<AuthRepository> { AuthRepositoryImpl(get()) }
-    }
