--- auth/domain/src/main/kotlin/com/filantrop/pvnclient/auth/domain/AuthInteractory.kt
@@ -1,29 +1,0 @@
-package com.filantrop.pvnclient.auth.domain
-
-import com.filantrop.pvnclient.core.model.UserData
-import kotlinx.coroutines.flow.Flow
-import kotlinx.coroutines.flow.map
-import org.koin.core.annotation.Single
-
-interface AuthInteractor {
-    val userData: Flow<UserData>
-
-    suspend fun loginWithToken(token: String)
-
-    suspend fun logout()
-}
-
-@Single(binds = [AuthInteractor::class])
-class AuthInteractorImpl(
-    private val authRepository: AuthRepository,
-) : AuthInteractor {
-    override val userData: Flow<UserData> =
-        authRepository.token.map {
-            requireNotNull(it) { "Token is null" }
-            UserData(it)
-        }
-
-    override suspend fun loginWithToken(token: String) = authRepository.loginWithToken(token)
-
-    override suspend fun logout() = authRepository.logout()
-}
