--- auth/ui/src/main/kotlin/com/filantrop/pvnclient/auth/ui/AuthScreen.kt
@@ -1,6 +1,0 @@
-package com.filantrop.pvnclient.auth.ui
-
-import androidx.compose.runtime.Composable
-
-@Composable
-fun AuthScreen() { /* no-op */ }
