+++ buildSrc/build.gradle.kts
@@ -12,15 +12,11 @@ gradlePlugin {
     plugins {
         register("androidApplication") {
             id = "pvnclient.android.application"
-            implementationClass = "com.filantrop.pvnclient.gradle.AndroidApplicationConventionPlugin"
-        }
-        register("androidLibrary") {
-            id = "pvnclient.android.library.android"
-            implementationClass = "com.filantrop.pvnclient.gradle.AndroidLibraryConventionPlugin"
+            implementationClass = "AndroidApplicationConventionPlugin"
         }
         register("kotlinLibrary") {
             id = "pvnclient.android.library.kotlin"
-            implementationClass = "com.filantrop.pvnclient.gradle.KotlinLibraryConventionPlugin"
+            implementationClass = "KotlinLibraryConventionPlugin"
         }
     }
 }
@@ -31,4 +27,5 @@ dependencies {
     implementation(libs.android.gradle.plugin)
     implementation(libs.guava)
     implementation(libs.kotlin.gradle.plugin)
+    implementation(libs.ksp.gradle.plugin)
 }
