+++ buildSrc/src/main/kotlin/KotlinLibraryConventionPlugin.kt
@@ -1,0 +1,16 @@
+import org.gradle.api.Plugin
+import org.gradle.api.Project
+import org.gradle.kotlin.dsl.dependencies
+
+class KotlinLibraryConventionPlugin : Plugin<Project> {
+    override fun apply(target: Project) {
+        with(target) {
+            with(pluginManager) {
+                apply("kotlin")
+                apply("com.autonomousapps.dependency-analysis")
+            }
+            dependencies {
+            }
+        }
+    }
+}
