--- buildSrc/src/main/kotlin/com/filantrop/pvnclient/gradle/KotlinLibraryConventionPlugin.kt
@@ -1,18 +1,0 @@
-package com.filantrop.pvnclient.gradle
-
-import org.gradle.api.Plugin
-import org.gradle.api.Project
-import org.gradle.kotlin.dsl.dependencies
-
-class KotlinLibraryConventionPlugin : Plugin<Project> {
-    override fun apply(target: Project) {
-        with(target) {
-            with(pluginManager) {
-                apply("kotlin")
-                apply("com.autonomousapps.dependency-analysis")
-            }
-            dependencies {
-            }
-        }
-    }
-}
