+++ config/detekt/baseline.xml
@@ -2,5 +2,7 @@
 <SmellBaseline>
   <ManuallySuppressedIssues></ManuallySuppressedIssues>
   <CurrentIssues>
+    <ID>MissingPackageDeclaration:AndroidApplicationConventionPlugin.kt$.AndroidApplicationConventionPlugin.kt</ID>
+    <ID>MissingPackageDeclaration:KotlinLibraryConventionPlugin.kt$.KotlinLibraryConventionPlugin.kt</ID>
   </CurrentIssues>
 </SmellBaseline>
