--- core/model/build.gradle.kts
@@ -1,6 +1,0 @@
-plugins {
-    id("pvnclient.android.library.kotlin")
-}
-
-dependencies {
-}
