--- core/model/src/main/kotlin/com/filantrop/pvnclient/core/model/UserData.kt
@@ -1,5 +1,0 @@
-package com.filantrop.pvnclient.core.model
-
-data class UserData(
-    val token: String,
-)
