--- core/persistent/build.gradle.kts
@@ -1,19 +1,0 @@
-import com.filantrop.pvnclient.gradle.extensions.ksp
-
-plugins {
-    id("pvnclient.android.library.android")
-    id("com.google.devtools.ksp")
-}
-
-android {
-    namespace = "com.filantrop.pvnclient.core.persistent"
-}
-
-dependencies {
-
-    implementation(libs.datastore)
-    implementation(libs.koin.annotations.jvm)
-    implementation(libs.koin.core)
-
-    ksp(libs.koin.ksp.compiler)
-}
