--- core/persistent/src/main/kotlin/com/filantrop/pvnclient/core/persistent/di/PersistentModule.kt
@@ -1,10 +1,0 @@
-package com.filantrop.pvnclient.core.persistent.di
-
-import com.filantrop.pvnclient.core.persistent.PreferenceStore
-import com.filantrop.pvnclient.core.persistent.PreferencesStoreImpl
-import org.koin.dsl.module
-
-val persistentModule =
-    module {
-        single<PreferenceStore> { PreferencesStoreImpl(get()) }
-    }
