+++ gradle/libs.versions.toml
@@ -1,55 +1,42 @@
 [versions]
-activity = "1.10.0"
+activity = "1.9.3"
 appcompat = "1.7.0"
 autonomousappsSorting = "0.14"
 autonomousappsUnused = "2.7.0"
-compose = "1.7.6"
-composeMaterial3 = "1.3.1"
 constraintlayout = "2.2.0"
-coroutinesCore = "1.10.1"
-datastore = "1.1.2"
 detektPlugin = "1.23.7"
 gradlePlugin = "8.2.2"
 guava = "33.4.0-jre"
 ipaddress = "5.5.1"
 java = "17"
 junit = "4.13.2"
-koin = "4.0.1"
-koinKsp = "2.0.0-Beta3"
-kotlin = "2.1.0"
-ksp = "2.1.0-1.0.29"
+junitTest = "1.2.1"
+kotlin = "1.9.24"
+ksp = "2.0.20-1.0.24"
 lombock = "1.18.36"
 material = "1.12.0"
+monitor = "1.7.2"
 okhttp = "4.12.0"
 protobuf = "4.26.1"
 protobufPlugin = "0.9.4"
 roomRuntime = "2.6.1"
-splashScreen = "1.0.1"
 
 [libraries]
 android-gradle-plugin = { group = "com.android.tools.build", name = "gradle", version.ref = "gradlePlugin" }
 androidx-activity = { group = "androidx.activity", name = "activity", version.ref = "activity" }
 androidx-appcompat = { group = "androidx.appcompat", name = "appcompat", version.ref = "appcompat" }
-androidx-compose-material3 = { group = "androidx.compose.material3", name = "material3", version.ref = "composeMaterial3" }
-androidx-compose-runtime-android = { group = "androidx.compose.runtime", name = "runtime-android", version.ref = "compose" }
-androidx-compose-ui = { group = "androidx.compose.ui", name = "ui", version.ref = "compose" }
 androidx-constraintlayout = { group = "androidx.constraintlayout", name = "constraintlayout", version.ref = "constraintlayout" }
-androidx-core-splashscreen = { group = "androidx.core", name = "core-splashscreen", version.ref = "splashScreen" }
+androidx-junit = { group = "androidx.test.ext", name = "junit", version.ref = "junitTest" }
+androidx-monitor = { group = "androidx.test", name = "monitor", version.ref = "monitor" }
 androidx-room-compiler = { module = "androidx.room:room-compiler", version.ref = "roomRuntime" }
+androidx-room-guava = { module = "androidx.room:room-guava", version.ref = "roomRuntime" }
 androidx-room-runtime = { module = "androidx.room:room-runtime", version.ref = "roomRuntime" }
-datastore = { group = "androidx.datastore", name = "datastore-preferences", version.ref = "datastore" }
 detekt = { group = "io.gitlab.arturbosch.detekt", name = "detekt-gradle-plugin", version.ref = "detektPlugin" }
 guava = { module = "com.google.guava:guava", version.ref = "guava" }
 ipaddress = { module = "com.github.seancfoley:ipaddress", version.ref = "ipaddress" }
 junit = { group = "junit", name = "junit", version.ref = "junit" }
-koin-android = { group = "io.insert-koin", name = "koin-android", version.ref = "koin" }
-koin-androidx-compose = { group = "io.insert-koin", name = "koin-androidx-compose", version.ref = "koin" }
-koin-annotations-jvm = { group = "io.insert-koin", name = "koin-annotations-jvm", version.ref = "koinKsp" }
-koin-core = { group = "io.insert-koin", name = "koin-core", version.ref = "koin" }
-koin-ksp-compiler = { group = "io.insert-koin", name = "koin-ksp-compiler", version.ref = "koinKsp" }
-koin-test-jvm = { group = "io.insert-koin", name = "koin-test-jvm", version.ref = "koin" }
 kotlin-gradle-plugin = { group = "org.jetbrains.kotlin", name = "kotlin-gradle-plugin", version.ref = "kotlin" }
-kotlinx-coroutines-core = { group = "org.jetbrains.kotlinx", name = "kotlinx-coroutines-core", version.ref = "coroutinesCore" }
+ksp-gradle-plugin = { group = "com.google.devtools.ksp", name = "com.google.devtools.ksp.gradle.plugin", version.ref = "ksp" }
 lombock = { group = "org.projectlombok", name = "lombok", version.ref = "lombock" }
 material = { group = "com.google.android.material", name = "material", version.ref = "material" }
 okhttp = { group = "com.squareup.okhttp3", name = "okhttp", version.ref = "okhttp" }
@@ -59,5 +46,5 @@ protoc = { module = "com.google.protobuf:protoc", version.ref = "protobuf" }
 [plugins]
 deps-sorting = { id = "com.squareup.sort-dependencies", version.ref = "autonomousappsSorting" }
 deps-unused = { id = "com.autonomousapps.dependency-analysis", version.ref = "autonomousappsUnused" }
-ksp = { id = "com.google.devtools.ksp", version.ref = "ksp" }
 protobuf = { id = "com.google.protobuf", version.ref = "protobufPlugin" }
+
